Remove old API client and rename the service ID for the new one (#20)
