+++ src/DependencyInjection/WebgriffeSyliusAkeneoExtension.php
@@ -55,7 +55,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
             'arguments' => [
                 'sylius.factory.product_image',
                 'sylius.repository.product_image',
-                'webgriffe_sylius_akeneo.api_client_official',
+                'webgriffe_sylius_akeneo.api_client',
             ],
         ],
         'immutable_slug' => [
@@ -70,7 +70,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
         'product_option' => [
             'class' => ProductOptionValueHandler::class,
             'arguments' => [
-                'webgriffe_sylius_akeneo.api_client_official',
+                'webgriffe_sylius_akeneo.api_client',
                 'sylius.repository.product_option',
                 'sylius.factory.product_option_value',
                 'sylius.factory.product_option_value_translation',
@@ -97,7 +97,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
         'file_attribute' => [
             'class' => FileAttributeValueHandler::class,
             'arguments' => [
-                'webgriffe_sylius_akeneo.api_client_official',
+                'webgriffe_sylius_akeneo.api_client',
                 'filesystem',
             ],
         ],
@@ -125,7 +125,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
             'arguments' => [
                 '$productImageFactory' => 'sylius.factory.product_image',
                 '$productImageRepository' => 'sylius.repository.product_image',
-                '$apiClient' => 'webgriffe_sylius_akeneo.api_client_official',
+                '$apiClient' => 'webgriffe_sylius_akeneo.api_client',
             ],
         ],
         'immutable_slug' => [
@@ -140,7 +140,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
         'product_option' => [
             'class' => ProductOptionValueHandler::class,
             'arguments' => [
-                '$apiClient' => 'webgriffe_sylius_akeneo.api_client_official',
+                '$apiClient' => 'webgriffe_sylius_akeneo.api_client',
                 '$productOptionRepository' => 'sylius.repository.product_option',
                 '$productOptionValueFactory' => 'sylius.factory.product_option_value',
                 '$productOptionValueTranslationFactory' => 'sylius.factory.product_option_value_translation',
@@ -169,7 +169,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
         'file_attribute' => [
             'class' => FileAttributeValueHandler::class,
             'arguments' => [
-                '$apiClient' => 'webgriffe_sylius_akeneo.api_client_official',
+                '$apiClient' => 'webgriffe_sylius_akeneo.api_client',
                 '$filesystem' => 'filesystem',
             ],
         ],
