+++ tests/Behat/Context/Setup/AkeneoContext.php
@@ -7,13 +7,13 @@ namespace Tests\Webgriffe\SyliusAkeneoPlugin\Behat\Context\Setup;
 use Akeneo\Pim\ApiClient\AkeneoPimClientInterface;
 use Behat\Behat\Context\Context;
 use DateTime;
-use Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\OfficialApiClientMock;
+use Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\ApiClientMock;
 use Webmozart\Assert\Assert;
 
 final class AkeneoContext implements Context
 {
     /**
-     * @param AkeneoPimClientInterface|OfficialApiClientMock $apiClient
+     * @param AkeneoPimClientInterface|ApiClientMock $apiClient
      */
     public function __construct(
         private $apiClient
@@ -25,7 +25,7 @@ final class AkeneoContext implements Context
      */
     public function thereIsAProductUpdatedAtOnAkeneo(string $identifier, DateTime $date): void
     {
-        Assert::isInstanceOf($this->apiClient, OfficialApiClientMock::class);
+        Assert::isInstanceOf($this->apiClient, ApiClientMock::class);
         $this->apiClient->addProductUpdatedAt($identifier, $date);
     }
 
@@ -35,7 +35,7 @@ final class AkeneoContext implements Context
      */
     public function thereAreProductsOnAkeneo(int $count): void
     {
-        Assert::isInstanceOf($this->apiClient, OfficialApiClientMock::class);
+        Assert::isInstanceOf($this->apiClient, ApiClientMock::class);
         for ($i = 1; $i <= $count; ++$i) {
             $this->apiClient->addProductUpdatedAt('product-' . $i, new DateTime());
         }
