+++ tests/Integration/TestDouble/ProductApiMock.php
@@ -33,7 +33,7 @@ final class ProductApiMock implements ProductApiInterface
 
     public function get(string $code, array $queryParameters = []): array
     {
-        return OfficialApiClientMock::jsonFileOrHttpNotFoundException(
+        return ApiClientMock::jsonFileOrHttpNotFoundException(
             __DIR__ . '/../DataFixtures/ApiClientMock/Product/' . $code . '.json'
         );
     }
